minor code refactoring to reduce LoC on views.py
